amigo/syntax: read source into a string upfront
The original design of the syntax scanner was to minimize how much of
the source file was kept in memory at a time. But I'm not convinced
this was a good trade off: it's much simpler to just read the full
source files into memory as a string, and then just return substrings.