amigo/syntax: simplify source position tracking
The prior semantics for source.r were that it's the buffer index after
ch, but it turns out to be much more useful if it's the buffer index
of ch itself.

While here, we can replace source.col (the column byte offset of ch)
with source.bol (the buffer index of the first character of the
current line). This fixes what I believe to be a column-tracking
desync issue in source.rewind (only affects syntactically invalid
files containing ".."), and is (at least IMO) much easier to reason
about.