amigo/syntax: add forward/backward links into tape
This commit extends the token tape to keep forward/backward links
between opening and closing punctuation tokens. Eventually, I
anticipate this to be useful for random-access / on-demand parsing of
the tape.

For now, I've just scattered a bunch of validation check into the
parser to check that links match up as expected, at least for
syntactically valid source files.