amigo/syntax: actual out-of-order parsing
This commit changes the parser to actually start parsing things out of
order, where possible.

There are several places where we still have to synchronously parse,
because of limitations of the current AST representation or internal
parser APIs. These have been left as asyncTODO with comments
explaining what needs to be done.