amigo/types: separate building functions from calling init functions
The previous code claimed to not care about the order that
buildFuncDecl is called, but actually it *was* necessary to ensure
that init functions are called in the correct order.

This commit splits the buildFuncDecl loop into two separate loops: one
just for compiling the function declarations, and a separate for
creating calls to the init functions.

This is a step towards being able to merge buildFunction into the
function body type checking code (i.e., hypothesis 4).