amigo/syntax: better PragmaHandler API
Currently, PragmaHandler is a single function overloaded to mean
several things, and the parser is responsible for maintaining its
state. Instead, we can just make it an interface, and allow the
handler to manage its own state.

While here, fix issues around handling of `blank`. In particular, the
`s.tapeelem = tapeelem{}` line meant to clear unneeded fields was
actually also clearing the previous position information, which the
scanner needs for detecting whether the `//go:` directive is the first
thing to appear on the line.

Also, when manually adding tapeelems because of an error handler
callback, we need to make sure to manually save `old.blank` too.